+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/bootstrap.py
@@ -15,6 +15,7 @@
 import argparse
 import logging
 import sys
+from importlib.metadata import PackageNotFoundError, distribution
 from subprocess import (
     PIPE,
     CalledProcessError,
@@ -23,7 +24,7 @@ from subprocess import (
     check_call,
 )
 
-import pkg_resources
+from packaging.requirements import Requirement
 
 from opentelemetry.instrumentation.bootstrap_gen import (
     default_instrumentations,
@@ -94,15 +95,18 @@ def _is_installed(req):
     if req in sys.modules:
         return True
 
+    req = Requirement(dep)
+
     try:
-        pkg_resources.get_distribution(req)
-    except pkg_resources.DistributionNotFound:
+        dist = distribution(req.name)
+    except PackageNotFoundError:
         return False
-    except pkg_resources.VersionConflict as exc:
+
+    if not req.specifier.contains(dist.version):
         logger.warning(
             "instrumentation for package %s is available but version %s is installed. Skipping.",
-            exc.req,
-            exc.dist.as_requirement(),  # pylint: disable=no-member
+            req,
+            dist.version,
         )
         return False
     return True