+++ opentelemetry-instrumentation/tests/test_distro.py
@@ -15,7 +15,7 @@
 
 from unittest import TestCase
 
-from pkg_resources import EntryPoint
+from importlib_metadata import EntryPoint
 
 from opentelemetry.instrumentation.distro import BaseDistro
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
@@ -34,7 +34,8 @@ class MockInstrumetor(BaseInstrumentor):
 
 class MockEntryPoint(EntryPoint):
     def __init__(self, obj):  # pylint: disable=super-init-not-called
-        self._obj = obj
+        # EntryPoint is immutable and prevents `self._obj = obj`
+        vars(self).update(_obj=obj)
 
     def load(self, *args, **kwargs):  # pylint: disable=signature-differs
         return self._obj